Consitent symmetric vertex notation
The assumption is to keep backward compatibility and do not rewrite
file format and converters used at pre-proessing step. Only neko
internals are changed and the points are swapped at the file reading
step. Periodic bc, dof numberring and grid generation are corrected
as well.
Stuff left to change/test:
- mesh file writers (e.g. nmsh_file_write) have to swap points
- curved faces/edges are not tested yet
- vtk file writers
- krylov/pc_hsmg.f90 should be checked as it operates on elements