Updated re2 file reader to support version 4.
This version of re2 was introduced into Nek5000 in September 2023.

Header format of the re2 file was changed in version 4.
Updates have been made to re2_file.f90 to support the new header in addition to the old.

Additionally added support for EXO boundary types. This allows "sidesets" from exodus to be used for setting boundary conditions in the case files.