Add user_startup routine, make case public in neko_simcomps (#1774)
The PR contains some improvements for user-file lovers.

- I make the neko_simcomps%case object public again, so that the user
file is in "god mode" via it.
- I add docstrings to the user file .f90 and do a bit of refactoring and
reformatting.
- A new user routine is added, called `user_startup`. This is run right
at the beginning of the case init and only gets the case file JSON as
input. The purpose is to either hack into the params before they are
passed to constructors, or define user-defined variables of built-in
type (constants in particular). A case in point is the material property
constants in the rayleigh and RBC examples. I guess this will be a
welcome addition. What prompted me to add this is that in the PR with
scalar improvements, the material_properties routine no longer received
the full case params, so doing things as before in the examples above
did not work.

---------

Co-authored-by: Victor Baconnet <baconnet@kth.se>
Co-authored-by: Tim Felle Olsen <tife@dtu.dk>