Add option for fluid statistics filename (#1723)
There was a suggestion to have the option to be able to get the same
behavior as before for the stats.

With this it is possible to replicate the original behavior if one wants
to by setting the filename to "fluid_stats.fld".

One can also set some other name if they want to.

Maybe we should add this keyword to the top of the simulation components
at some point as well.