See if we can simplify the ksp factory (#1762)
See issue #1313

* I introduce a deferred init for ksp_t. Some constructors had `target`
on this, and some didn't. I chose to have `target` in the shared
interface.
* Remove the `krylov_solver_destroy` subroutine. We should be able to
just call %free on the object.
* Remove the `select` type in the factory and use the shared `init`.


Caveat:
GMRES and CACG had additional paramters to their init, which had to be
removed :(. I just set the default values in the type definition. The
code never used any values but the defaults though, since I didn't have
to change a single constructor call.
We could add additional constructors for those type, which account for
the extra paramters, if we want to.