Make chkp not output at time 0 (#1797)
As pointed out by @pschlatt1 my fix for start_time to avoid writing
chkps at time 0 did not work, they only got outputted at the second time
step instead 🙃

This is a fix for that.

A note is that working with the output is quite confusing and in the
future we should give the counter, nexecution logic some love.