Remove built-in zones from the code (#1791)
I think we are ready, as our tooling does not allow them to appear
(rea2nbin creates labled zones from built-in ones).

I retain the (now strange) numbering that periodic zones are marked with
5 and labeled zones with 7 in the mesh file. Changing this would cause
too much disruption for existing meshes, so we can save that for when
and if we decide to revamp the nmsh format.

---------

Co-authored-by: Tim Felle Olsen <timfelle@hotmail.com>