Only execute field_dirichlet_update() once per timestep (#1792)
The fact that we apply boundary conditions twice per timestep (see
[here](https://github.com/ExtremeFLOW/neko/blob/24660f35a432d43dc911371585531c41c0375977/src/fluid/fluid_scheme_incompressible.f90#L513-L541))
can cause unwanted behaviors with `field_dirichlet` and
`field_dirichlet_vector` when calling the `field_dirichlet_update`
subroutine.

Specifically, it adds extra overhead when, for example, there is reading
of files involved or some time-dependent computation at every timestep,
which is then done twice per timestep and can add redundant overhead.
Technically, **all** BCs are applied twice, but field_dirichlet struck
me as the most critical one since user stuff is involved.

I don't know if this also applies to the `wall_model_bc` since it has a
`compute` and an `apply`? I have no idea if that is generally an
expensive computation or not @timofeymukha.

The proposed solution involves adding an attribute `tstep_applied` to
the `field_dirichlet_t` and `field_dirichlet_vector_t` objects, which
stores the value of the time step it has latest called `this%update()`
at. This logic is pretty simple and at quite a low level in the
`field_dirichlet` objects, but it can serve as basis for discussions
(hence the draft).