Add only_facets to bc finalize interface. (#1824)
Adds `only_facets` to the `finalize` interface of `bc_t`, so that it can
be passed to `init_base`.
Needed, since e.g. wall-models need to ensure `only_facets=.true.`. 

The default value for all bcs is set based on what is the current (and
correct) behavior.

Bonus: the wall_model bc now writes a separate field called
sampling_height, which is quite useful for debugging the case setup.

---------

Co-authored-by: Tim Felle Olsen <timfelle@hotmail.com>