Constructors for simulation_component_t without JSON dep (#1820)
Depends on #1819 

This PR addresses an architectural issue with simcomps: the constructor
of the base type depends on JSON. So there really is no way to init a
simcomp from components. Even though we have some constructors from
components in simcomps, they don't fully work properly, because
`init_base` is always run in the json-based constuctor.

Here, I add two additional constructors to the base simcomp type, which
are based on components. A `generic` called (predictably)
`init_from_components` binds them together.

The individual simcomps need to make use of this. So far in this PR, I
just modify the `field_writer` simcomp, but this change needs to later
propagate to the other simcomps, and be adopted in new ones. When done,
we will be able to init simcomps without json inside neko, the user
file, or 3rd-party libraries like Tim's.