add start_time for probes (#1846)
As the title says, adds a `start_time` to the probes. It is basically
the same behavior as the statistics, with the intent to skip any
transient behaviors in the beginning of a simulation, and start
collecting probes after the given `start_time`.

It would be cool to have this stuff at the simcomp level but I am not
feeling brave enough to mess with the time controllers right now...

I propose to merge this if we are happy to get the functionality, and
think bigger picture later :)