Fix bounds in masked gather/scatter copies (#1860)
I think this is correct?

We should always loop across the size of the contiguous array. For the
gather mask, this is `a`, so we loop until `m`. For the scatter mask,
this is `b`, so we loop until `n`. We assume that the array that is
addressed via the mask is at least the same size. The size of the mask
is also the same as that of the contiguous array.