Redesign material_properties user routine and their handling in schemes (#1779)
The following changes are introduces to the material properties handling
- No components in `_scheme` for constant values, both scalar and fluid
have `field_t` for material properties.
- A helper `material_properties` field list is used to collect pointers
to all the material properties of a scheme together. Used in the user
routine, see below.
- The `pnpn` schemes currently hard-code using a constant `rho` and it
is also used in the `rhs_maker`. Addressing the latter is left for a
future PR, currently we just pass e.g. `this%rho%x(1,1,1,1)`. This is
not different from what we had before in terms of functionality.
- The schemes call their own update routines for material properties,
which calls the user routine, and, when needed, adds the turbulent
component to diffusivity.
- simulation.f90 is cleaned up.
- The density of the scalar is changed to be a pointer to a field, now
pointing to the fluid, so updates in the fluid density are respected by
the scalar.
- A new keyword for fluid is introduced: `full_stress_formulation`,
which changes the governing equations to have the full form of the
viscous stress tensor. False by default.

The following changes are introduced to the user material_properties
scheme
* Adds `name` to distinguish between callees. The fluid and scalar send
their `name` component.
* A properties field list is received and manipulated. Can be of
arbitrary size in case new `schemes` use it. For incompressible fluid,
the first field is density and the second is mu. For the scalar, the
first is cp, and the second lambda.


Finally, the rayleigh_bendard and rayleigh_benard_cylinder examples are
updated to showcase the new stuff.

---------

Co-authored-by: Tim Felle Olsen <timfelle@hotmail.com>