Use a json subdict to extract and pass down (potential) precon params (#1851)
This add the functionallty of providing options to the preconditioner. A
use case is for example to specify a different coarse grid solver for
HSMG, tune the number of iterations and such.

The current implementation extract a subdict from the json file and pass
that down to precon init.

```json
        "velocity_solver": {
            "type": "cg",
            "preconditioner": {
                "type": "jacobi"
            },
            "projection_space_size": 4,
            "absolute_tolerance": 1e-7,
            "max_iterations": 800
        },
        "pressure_solver": {
            "type": "gmres",
            "preconditioner": {
                "type": "hsmg",
                "coarse_grid":  {
                    "solver": "cheby",
                    "iterations": 50
                },
            },
            "projection_space_size": 4,
            "absolute_tolerance": 1e-4,
            "max_iterations": 800
        },
```

This is a *breaking change* all case files needs to be updated to
include at least `preconditioner.type`


fixes #1849

---------

Co-authored-by: Tim Felle Olsen <timfelle@hotmail.com>