Add possibility to install with pixi (#1880)
Adds a pixi.toml for installing Neko with pixi. For now, CPU only,
though I think we can extend this later, since conda-forge can install
cuda and rocm I believe.

I think this is now the fastest and easiest way to get a CPU build with
all deps, and also a convenient isolated dev environment.
Please try it out on your machines as per the instructions.

---------

Co-authored-by: Niclas Jansson <njansson@kth.se>