Programming examples (#1882)
I've started preparing some programming examples in user files not
attached to a particular case. I wanted to have a .f90 as the format so
that one can actually check that they compile as the code gets updated.
I hope you like the format.

@Mr-2c You have requested something like this a long while back, I think
it would be good to have for v1.0. What topics would you think be good
to cover?

---------

Co-authored-by: Tim Felle Olsen <timfelle@hotmail.com>
Co-authored-by: Tim Felle Olsen <tife@dtu.dk>