Add a device_memset before any gs operations (for extra safety) (#1957)
Add a memset after gs buffers are allocated. This is not needed for
normal gather-scatter operations since they operate on the full length
(m) for each call, however async. operations (as the ones done by global
interpolation #1542, could be different).

This is similar to what was done in #1932, but here we use memset in the
operating stream, thus reduce the numbers of global sync. points.