Add  material properties to the field registry (#1906)
Added as `this%name`, plus `_rho`, `_mu`, `_lambda`, `_cp`.

I am mainly doing this to have easy access to this in wall models and
also force_torque simcomp.

I guess we are kid of drifting towards registering more and more things,
but maybe that's not too bad as long as the naming is solidly unique.