Integration tests with pytest (#1526)
I have now developed this into something that is mergable. For now,
parallel to reframe and the simple regression test. However, this
deprecates the smoke tests in reframe, so I commented it out of the CI,
at least for now.

So, this adds a pytest-based integration test framework, which can be
run on any machine (I think), provided a launcher scripts to start neko
for the tested environment. Please read the README.md for details.

I add two tests

- Smoke-screen example executions for tgv, hemi and rayleigh_bendard. We
can easily add any other example.
- Compilation of all the example user files with makeneko.

There is also a dummy test that just demonstrates how to add a test.

We will pass over the regression test with the cylinder we have to this
frameworks, once we have the updated log format.

---------

Co-authored-by: Tim Felle Olsen <tife@dtu.dk>
(cherry picked from commit 722802dffa51a7eaebbf92ac89a7fee1a44f9966)