Fix zero-sized array c_loc (#1976)
Fixes an issue when using a device backend, and the pnpn pressure
surface terms are zero since no symmetry conditions are present. In this
case, no facets will be marked, and `facet_normal_t` will, during init
map a zero sized host array to the device, causing an invalid zero-sized
`c_loc` call.

To reproduce the issue, run e.g. the `turb_pipe` example

This is probably the cause of some of the face_normal related issues we
have been observing with the device backend