Pass time_state_t to boundary condition apply routines (#2025)
In line with the same changes in other routines.

I mainly did this now for the sake of updating the user Dirichlet
routine to a similar pattern as in the #2021, which is done here as
well.

Killed quite a bit if(present) stuff, which is nice

---------

Co-authored-by: Tim Felle Olsen <timfelle@hotmail.com>
(cherry picked from commit a783d7a6eaad0be3c2ddf40d022ed8a101bca1d0)