Checkpoint file counter in reader (#2077)
A few relatively deep changes, however, the impact should be quite
limited.

The file checker of the generic file can now work on a specific file
name. which made it useful for checking counter based files, without
modifying the class name.

The checkpoint files can have a name specified with the suffix.

The checkpoint counter can be used when reading files.

One possible situation of unexpected behaviour which might arrise here.
Say you set the checkpoint file name to be "file" but do not overwrite
files, and then specify to load from checkpoint "file.chkp", you will
actually load "file00000.chkp" instead of throwing an error. So a user
might be confused why that point is loaded and not the one with the
highest number.