User statistics (#1519)
User defined statistics. 

Picks a field from the field_registry and computes the temporal average.
Going to look through for documentation and such, but most things are in
place.

The ordering of the output fields is kind of all over the place. Maybe
best if everything would be output into the scalars?

Finishes the CTR porting adventure, interpolation next.

Based on #1494 


Update from Timofey:
Added a test, cleaned up, added docs. Fixed a bug in fld output. As
Martin mentions, it really is better to put everything into scalars,
probably. But for now this is fine.

---------

Co-authored-by: Timofey Mukha <timofey.mukha@protonmail.com>