tamg: add pairwise aggregation. random iteration through elements (#2135)
Add a pairwise aggregation scheme for amg.
Add a random permutation to iterate through elements for aggregation.

fix improper use of doxygen style comments

fix issue with minimum aggregates per level. Now truncates the amg
hierarchy to end on the level at which there are insufficient dofs to
aggregate on at least one rank, and uses that level as the coarsest
level.