Stabilized-convergence logging (#2153)
Add initial grace period for convergence warnings and checks.

This should help reduce superflous "Solver did not converged" messages
as the flow stabilizes during the initial part of the simulation.

Here we declare a flow "stabilized" on the first iteration for which the
solver converged for all components.

To facilitate programatic use, we also introduce a subroutine to reset
the value.