Fix/zeroing_scratch (#2199)
This PR forces users to specify if they want the field requested from a
scratch registry to be `clear`ed before it is passed to them.

This then also influences all places where we used the registry, meaning
we need to be careful that I have not messed up which are returned
zeroed and which are assigned by the user.

Please check carefully, this should be tested

(cherry picked from commit 731421d6543163ba7d98e55c6be56cd753756a4f)