Fix scalar auto name, throw on duplicate (#2205)
`write(field_name, '(A,I0)') 's_', i` caused a segfault when field_name
is of length zero. It seems, `write` does not automatically reallocate
to the necessary size, unlike assignment. I added a buffer of fixed size
to write to.

I also suggest removing the handling of duplicates and throw an error. I
think a duplicate is a red flag meaning that something went not like the
user wanted.