Add tests for statistics, revert fld writer behaviour, fix scalar_stats_output (#2212)
* Adds a `freeze` component for the scalar_scheme, similar to the fluid.
This is to allow setting synthetic for the scalar in `compute` in the
use file.
* This needed a bit of change in scalar_aux, it treats one KSP only,
passed in a loop from `scalars`, based on the status of `freeze`.
@timfelle the `stabalized` thing was working somehow on all scalars at
the same time, or? I am a bit unsure if I broke something here.
* Adds tests for fluid and scalar *basic* stats. Outputs 1d, 2d and 3d
stats, checks the values against prescribed analytical ones.
* Fixes bug in scalar_stats_output: adds the reshuffling of fields for
consistent output.
* Adds mpi4py to the python environment. Hopefully this does not mess
things up?
* Updates the statistics guide with the new numbering which includes
temperature.