Add check for implicit integer conversion in json read (#2232)
json_fortran implicitly converts reals to integers when reading an entry
into an integer.

This adds another read to the same entry with a real variable and checks
that no rounding has occurred when populating the integer.

Also adds a skeleton for unit testing json_utils. Currently, just check
that integer reads do work when json contains reals that have a
0 decimal.