Fix mean_output_t initilization (#2235)
Asked AI why it thinks we get this weird error with user_stats.csv
failing to appear randomly in the pytest CI, and it directly found this
very sad bug.

`this%start_time` was not initialized, so randomly the check of time >
this%start_time would fail...

We should not have let that error slide like that.